requirements.txt
asgiref==3.6.0
bidict==0.22.1
certifi==2023.5.7
charset-normalizer==3.1.0
controller==0.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-admin-interface==0.26.0
django-admin-multi-select-filter==1.0.0
django-colorfield==0.8.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-environ==0.10.0
django-import-export==3.2.0
django-more-admin-filters==1.7
djangoql==0.17.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
groups==0.1
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
Pillow==9.5.0
ply==3.11
psycopg==3.1.9
psycopg2==2.9.6
PyJWT==2.7.0
python-engineio==4.4.1
python-slugify==8.0.1
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
repoze.lru==0.7
requests==2.30.0
Routes==2.5.1
six==1.16.0
sqlparse==0.4.4
tablib==3.4.0
text-unidecode==1.3
typing_extensions==4.5.0
urllib3==2.0.2
xlrd==2.0.1
xlwt==1.3.0
